requirements.txt
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.5.7
black==21.7b0
click==8.0.1
colorama==0.4.4
Django==3.2.6
django-crispy-forms==1.13.0
djangorestframework==3.12.4
drf-writable-nested==0.6.3
mypy-extensions==0.4.3
mysqlclient==2.0.3
pathspec==0.9.0
postgres==4.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
psycopg2-pool==1.1
pycodestyle==2.7.0
pytz==2021.1
regex==2021.8.3
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
